move Post.parse_markdown() to postutils
